Move meals from components to container
