+++ src/containers/Meals.js
@@ -2,10 +2,10 @@ import { Grid, Typography } from '@material-ui/core';
 import React from 'react';
 import { useSelector } from 'react-redux';
 import commonStyles from '../utils/commonStyles';
-import CircularProgressWithLabel from './CircularProgressWithLabel';
-import FoodKind from './FoodKind';
-import Header from './Header';
-import NavigationBar from './NavigationBar';
+import CircularProgressWithLabel from '../components/CircularProgressWithLabel';
+import FoodKind from '../components/FoodKind';
+import Header from '../components/Header';
+import NavigationBar from '../components/NavigationBar';
 import carbohydrates from '../assets/images/carbohydrates.cms';
 import proteins from '../assets/images/proteins.png';
 import fats from '../assets/images/fats.jpg';
